Stop registering run configuration producers; build configs directly in run actions
